Add authentication using REST API
