+++ backend/api-mock.yml
@@ -3,6 +3,7 @@ version: '3.7'
 services:
   prism:
     image: stoplight/prism:3
+    restart: always
     ports:
       - 9080:9080
     volumes:
