+++ src/App.tsx
@@ -3,10 +3,11 @@ import { NavigationContainer } from '@react-navigation/native';
 import { activateKeepAwake } from 'expo-keep-awake';
 import React from 'react';
 
-import Auth from './components/pages/Auth';
 import Home from './components/pages/Home';
+import Login from './components/pages/Login';
 import Picking from './components/pages/Picking';
 import { TrashContextProvider } from './contexts/TrashContext';
+import { UserContextProvider } from './contexts/UserContext';
 
 const Tab = createBottomTabNavigator();
 
@@ -15,15 +16,17 @@ const App: React.FC = () => {
     activateKeepAwake();
   }
   return (
-    <TrashContextProvider>
-      <NavigationContainer>
-        <Tab.Navigator initialRouteName="home">
-          <Tab.Screen {...Home} />
-          <Tab.Screen {...Picking} />
-          <Tab.Screen {...Auth} />
-        </Tab.Navigator>
-      </NavigationContainer>
-    </TrashContextProvider>
+    <UserContextProvider>
+      <TrashContextProvider>
+        <NavigationContainer>
+          <Tab.Navigator initialRouteName="home">
+            <Tab.Screen {...Home} />
+            <Tab.Screen {...Picking} />
+            <Tab.Screen {...Login} />
+          </Tab.Navigator>
+        </NavigationContainer>
+      </TrashContextProvider>
+    </UserContextProvider>
   );
 };
